requirements.txt
# Core video processing
opencv-python>=4.8.0
pytesseract>=0.3.10
easyocr>=1.7.0

# Video download
pytubefix>=8.12.1

# Data processing and analysis
numpy>=1.21.0
pandas>=1.5.0
scipy>=1.9.0
matplotlib>=3.5.0

# Image processing and hashing
imagehash>=4.3.0
Pillow>=9.0.0

# Progress bars and utilities
tqdm>=4.64.0

# Excel file support
openpyxl>=3.1.0